Split the unit tests into separate suites
