+++ mu-plugins/10up-plugin/composer.json
@@ -11,6 +11,13 @@
     "php": ">=7.0",
     "spatie/php-structure-discoverer": "^2.1"
   },
+  "require-dev": {
+    "phpunit/phpunit": "^9.5",
+    "yoast/phpunit-polyfills": "^2.0"
+  },
+  "scripts": {
+    "test": "./vendor/bin/phpunit"
+  },
   "autoload": {
     "psr-4": {
       "TenUpPlugin\\": "includes/classes/"
