+++ mu-plugins/10up-plugin/includes/classes/class-loader-cache/discoverer-cache-TenUpPlugin
@@ -1,0 +1,1 @@
+a:9:{i:0;s:38:"TenUpPlugin\PostTypes\AbstractPostType";i:1;s:26:"TenUpPlugin\PostTypes\Post";i:2;s:26:"TenUpPlugin\PostTypes\Demo";i:3;s:26:"TenUpPlugin\PostTypes\Page";i:4;s:42:"TenUpPlugin\PostTypes\AbstractCorePostType";i:5;s:18:"TenUpPlugin\Module";i:6;s:39:"TenUpPlugin\Taxonomies\AbstractTaxonomy";i:7;s:27:"TenUpPlugin\Taxonomies\Demo";i:8;s:32:"TenUpPlugin\ModuleInitialization";}
