+++ themes/10up-theme/includes/classes/class-loader-cache/discoverer-cache-TenUpTheme
@@ -1,0 +1,1 @@
+a:2:{i:0;s:17:"TenUpTheme\Module";i:1;s:31:"TenUpTheme\ModuleInitialization";}
